Added narrow format and ability to add custom link
The search box now transforms if it is in a narrow space (380px or narrower, but customizable) or set it to a high number (9999) and experience an alternate style all the time. Also added the ability to create a custom link to display under the search box with the other links. Next up: Multi facet selection. (Winter 2017 maybe)